+++ demo/app/app.component.ts
@@ -525,6 +525,15 @@ function moo() {
     this.dialogMngr.create(options);
   }
 
+  openDialogAndPauseHotkeys(options) {
+    const dlg = this.dialogMngr.create(options);
+    this.hotkeysService.pauseOthers(dlg.instance);
+    dlg.instance.close.subscribe(() => {
+      console.log('close');
+      this.hotkeysService.unpauseOthers(dlg.instance);
+    });
+  }
+
   menuClicked(event) {
     console.log('Menu clicked', event);
   }